started adding some error handling and moving files into folders so file structure makes more sense
