+++ README.md
@@ -1,4 +1,4 @@
-# LocalLLM Plugin
+# LLM Plugin
 
 This is a sample plugin for Obsidian (https://obsidian.md).
 
@@ -94,6 +94,4 @@ If you have multiple URLs, you can also do:
 ## API Documentation
 
 See https://github.com/obsidianmd/obsidian-api
-# LocalLLMPlugin
-# LocalLLMPlugin
-# LocalLLMPlugin
+# LLMPlugin
