+++ src/Plugin/Widget/Widget.ts
@@ -2,15 +2,15 @@ import { ChatContainer } from "Plugin/Components/ChatContainer";
 import { Header } from "Plugin/Components/Header";
 import { HistoryContainer } from "Plugin/Components/HistoryContainer";
 import { SettingsContainer } from "Plugin/Components/SettingsContainer";
-import LocalLLMPlugin, { DEFAULT_SETTINGS } from "main";
+import LLMPlugin, { DEFAULT_SETTINGS } from "main";
 import { ItemView, WorkspaceLeaf } from "obsidian";
-import { classNames } from "utils/utils";
+import { classNames } from "utils/classNames";
 
 export const VIEW_TYPE = "example-view";
 
 export class WidgetView extends ItemView {
-	plugin: LocalLLMPlugin;
-	constructor(leaf: WorkspaceLeaf, plugin: LocalLLMPlugin) {
+	plugin: LLMPlugin;
+	constructor(leaf: WorkspaceLeaf, plugin: LLMPlugin) {
 		super(leaf);
 		this.plugin = plugin;
 	}
