+++ src/utils/classNames.ts
@@ -1,0 +1,24 @@
+
+export const classNames = {
+	modal: {
+		"messages-div": "modal-messages-div",
+		"title-border": "modal-title-border",
+		"prompt-container": "modal-prompt-container",
+		"text-area": "modal-chat-prompt-text-area",
+		button: "modal-send-button",
+	},
+	widget: {
+		"messages-div": "widget-messages-div",
+		"title-border": "widget-title-border",
+		"prompt-container": "widget-prompt-container",
+		"text-area": "widget-chat-prompt-text-area",
+		button: "widget-send-button",
+	},
+	"floating-action-button": {
+		"messages-div": "fab-messages-div",
+		"title-border": "fab-title-border",
+		"prompt-container": "fab-prompt-container",
+		"text-area": "fab-chat-prompt-text-area",
+		button: "fab-send-button",
+	},
+};
