+++ src/utils/utils.ts
@@ -6,7 +6,7 @@ import {
 	ViewType,
 } from "Types/types";
 import { existsSync } from "fs";
-import LocalLLMPlugin from "main";
+import LLMPlugin from "main";
 import { Editor } from "obsidian";
 import OpenAI from "openai";
 import { Stream } from "openai/streaming";
@@ -93,7 +93,7 @@ export function processReplacementTokens(prompt: string) {
 }
 
 export function getViewInfo(
-	plugin: LocalLLMPlugin,
+	plugin: LLMPlugin,
 	viewType: ViewType
 ): ViewSettings {
 	if (viewType === "modal") {
@@ -140,7 +140,7 @@ export function getViewInfo(
 }
 
 export function setHistoryIndex(
-	plugin: LocalLLMPlugin,
+	plugin: LLMPlugin,
 	viewType: ViewType,
 	length?: number
 ) {
@@ -180,156 +180,3 @@ export function appendMessage(editor: Editor, message: string, type?: string) {
 
 	moveCursorToEndOfFile(editor!);
 }
-
-export function serializeMessages(messages: Message[]) {
-	let response = "";
-	messages.forEach((message: Message) => {
-		response += `${message.role} : ${message.content}\n\n`;
-	});
-
-	return response;
-}
-
-export const models: Record<string, Model> = {
-	"Mistral OpenOrca": {
-		model: "mistral-7b-openorca.Q4_0.gguf",
-		type: "GPT4All",
-		endpoint: "chat",
-		url: "/v1/chat/completions",
-	},
-	"Mistral Instruct": {
-		model: "mistral-7b-instruct-v0.1.Q4_0.gguf",
-		type: "GPT4All",
-		endpoint: "chat",
-		url: "/v1/chat/completions",
-	},
-	"GPT4All Falcon": {
-		model: "gpt4all-falcon-newbpe-q4_0.gguf",
-		type: "GPT4All",
-		endpoint: "chat",
-		url: "/v1/chat/completions",
-	},
-	"Orca 2 (Medium)": {
-		model: "orca-2-7b.Q4_0.gguf",
-		type: "GPT4All",
-		endpoint: "chat",
-		url: "/v1/chat/completions",
-	},
-	"Orca 2 (Full)": {
-		model: "orca-2-13b.Q4_0.gguf",
-		type: "GPT4All",
-		endpoint: "chat",
-		url: "/v1/chat/completions",
-	},
-	"Mini Orca (Small)": {
-		model: "orca-mini-3b-gguf2-q4_0.gguf",
-		type: "GPT4All",
-		endpoint: "chat",
-		url: "/v1/chat/completions",
-	},
-	"MPT Chat": {
-		model: "mpt-7b-chat-newbpe-q4_0.gguf",
-		type: "GPT4All",
-		endpoint: "chat",
-		url: "/v1/chat/completions",
-	},
-	"Wizard v1.2": {
-		model: "wizardlm-13b-v1.2.Q4_0.gguf",
-		type: "GPT4All",
-		endpoint: "chat",
-		url: "/v1/chat/completions",
-	},
-	Hermes: {
-		model: "nous-hermes-llama2-13b.Q4_0.gguf",
-		type: "GPT4All",
-		endpoint: "chat",
-		url: "/v1/chat/completions",
-	},
-	Snoozy: {
-		model: "gpt4all-13b-snoozy-q4_0.gguf",
-		type: "GPT4All",
-		endpoint: "chat",
-		url: "/v1/chat/completions",
-	},
-	"EM German Mistral": {
-		model: "em_german_mistral_v01.Q4_0.gguf",
-		type: "GPT4All",
-		endpoint: "chat",
-		url: "/v1/chat/completions",
-	},
-	"ChatGPT-3.5 Turbo": {
-		model: "gpt-3.5-turbo",
-		type: "openAI",
-		endpoint: "chat",
-		url: "/chat/completions",
-	},
-	// "Text Embedding 3 (Small)": {
-	// 	model: "text-embedding-3-small",
-	// 	type: "openAI",
-	// 	endpoint: "embeddings",
-	// 	url: "/embeddings",
-	// },
-	// "DALL·E 3": {
-	// 	model: "dall-e-3",
-	// 	type: "openAI",
-	// 	endpoint: "images",
-	// 	url: "/images/generations",
-	// // },
-	// "DALL·E 2": {
-	// 	model: "dall-e-2",
-	// 	type: "openAI",
-	// 	endpoint: "images",
-	// 	url: "/images/generations",
-	// },
-};
-
-export const modelNames: Record<string, string> = {
-	"mistral-7b-openorca.Q4_0.gguf": "Mistral OpenOrca",
-	"mistral-7b-instruct-v0.1.Q4_0.gguf": "Mistral Instruct",
-	"gpt4all-falcon-newbpe-q4_0.gguf": "GPT4All Falcon",
-	"orca-2-7b.Q4_0.gguf": "Orca 2 (Medium)",
-	"orca-2-13b.Q4_0.gguf": "Orca 2 (Full)",
-	"orca-mini-3b-gguf2-q4_0.gguf": "Mini Orca (Small)",
-	"mpt-7b-chat-newbpe-q4_0.gguf": "MPT Chat",
-	"wizardlm-13b-v1.2.Q4_0.gguf": "Wizard v1.2",
-	"nous-hermes-llama2-13b.Q4_0.gguf": "Hermes",
-	"gpt4all-13b-snoozy-q4_0.gguf": "Snoozy",
-	"em_german_mistral_v01.Q4_0.gguf": "EM German Mistral",
-	"gpt-3.5-turbo": "ChatGPT-3.5 Turbo",
-	// "text-embedding-3-small": "Text Embedding 3 (Small)",
-	// "dall-e-3": "DALL·E 3",
-	// "dall-e-2": "DALL·E 2",
-};
-
-// export function getModelNames():Record<string, string> {
-//     let modelNames = Object.keys(models)
-//     let names:Record<string, string> = {}
-//     for(const name of modelNames) {
-//         names[models[name].model] = name
-//     }
-// 	return names
-// }
-
-export const classNames = {
-	modal: {
-		"messages-div": "modal-messages-div",
-		"title-border": "modal-title-border",
-		"prompt-container": "modal-prompt-container",
-		"text-area": "modal-chat-prompt-text-area",
-		button: "modal-send-button",
-	},
-	widget: {
-		"messages-div": "widget-messages-div",
-		"title-border": "widget-title-border",
-		"prompt-container": "widget-prompt-container",
-		"text-area": "widget-chat-prompt-text-area",
-		button: "widget-send-button",
-	},
-	"floating-action-button": {
-		"messages-div": "fab-messages-div",
-		"title-border": "fab-title-border",
-		"prompt-container": "fab-prompt-container",
-		"text-area": "fab-chat-prompt-text-area",
-		button: "fab-send-button",
-	},
-};
